finish node.js backend

+++ nodeBackEnd/services/playerService.ts
@@ -2,6 +2,7 @@ import { Prisma, PrismaClient } from "@prisma/client";
 import { Player } from "../Models/player";
 import { v4 as uuidv4 } from "uuid";
 import { info } from "console";
+import { PlayerDTOCreate, PlayerDTOUpdate } from "../Models/playerDTO";
 
 
 
@@ -21,18 +22,14 @@ export const getAllPlayersWithoutCountry = async () => {
     return await prisma.player.findMany()
 } 
 
-export const getPlayerById = async (playerId: string) => {
-    try {
-        const prisma = new PrismaClient()
-        const player = await prisma.player.findUnique({
-            include: { Country: true },
-            where: { id: playerId },
-        })
-        return player
-    } catch (err) {
-        console.log(err)
-    }
+export const getPlayerById = async (playerId: string) => {    
+    const prisma = new PrismaClient()
+    const player = await prisma.player.findUnique({
+        //include: { Country: true },
+        where: { id: `${playerId}` },
+    })
     
+    return player
 }
 
 export const getPercentageByCountry = async() => {
@@ -44,51 +41,66 @@ export const getPercentageByCountry = async() => {
     const percentages = new Map()
     players.map((player) => {
         if(percentages.has(player.Country?.code)) {
-            percentages.get(player.Country?.code).val++
+            let oldValue = percentages.get(player.Country?.code) as number
+            percentages.set(player.Country?.code, oldValue + 1)
         }
         else {
-            percentages.set(player.Country?.code, {val: 1})
+            percentages.set(player.Country?.code, 1)
         }
     })
 
     for (let entry of Array.from(percentages.entries())) {
-        percentages.set(entry[0], entry[1]/percentages.size * 100)
+        percentages.set(entry[0], entry[1]/players.length * 100)
     }
 
     return percentages
 }
 
-export const updatePlayer = async(player: Player) => {
+export const updatePlayer = async(player: PlayerDTOUpdate) => {
     const prisma = new PrismaClient()
-    return await prisma.player.update({
-        where: { id: player.id },
+    let updatedPlayer = await prisma.player.update({
+        where: { id: `${player.id}` },
         data: {
-            name: player.name,
+            name: `${player.name}`,
             active: player.active
         }
     })
+
+    return updatedPlayer !== null
 }
 
-export const createPlayer = async(player: Player) => {
+export const createPlayer = async(player: PlayerDTOCreate) => {
     const prisma = new PrismaClient()
     const existingPlayer = await prisma.player.findFirst({
         include: { Country: true },
         where: { 
             name: player.name,
             Country: {
-                code: player.country?.code
+                code: player.countryCode
             }
         }
     })
 
     if (existingPlayer === null) {
+        let newPlayer: Player = {
+            id: uuidv4(),
+            playerId: player.playerId,
+            name: player.name,
+            gender: player.gender,
+            active: player.active,
+            birthyear: player.birthyear,
+            playHand: player.playHand,
+            playStyle: player.playStyle,
+            grip: player.grip
+        }
+
         const existingCountry = await prisma.country.findFirst({
             where: {
-                code: player.country?.code
+                code: player.countryCode
             }
         })
 
-        let countryCode = player.country!.code
+        let countryCode = player.countryCode
         if (existingCountry === null) {
 
             let newCountry = await prisma.country.create({
@@ -97,27 +109,34 @@ export const createPlayer = async(player: Player) => {
                     code: countryCode
                 },
                 select: {
-                    id: true
+                    id: true,
+                    code: true
                 }
             })
 
-            player.country!.id = newCountry.id
+            newPlayer.country = {
+                id: newCountry.id,
+                code: newCountry.code ?? undefined
+            }
         } else {
-            player.country!.id = existingCountry.id
+            newPlayer.country = {
+                id: existingCountry.id,
+                code: existingCountry.code ?? undefined
+            }
         }
 
         return await prisma.player.create({
             data: {
-                id: uuidv4(),
-                playerId: player.playerId,
-                name: player.name,
-                gender: player.gender,
-                active: player.active,
-                birthyear: player.birthyear,
-                playHand: player.playHand,
-                playStyle: player.playStyle,
-                grip: player.grip,
-                countryId: player.country?.id
+                id: newPlayer.id!,
+                playerId: newPlayer.playerId,
+                name: newPlayer.name,
+                gender: newPlayer.gender,
+                active: newPlayer.active,
+                birthyear: newPlayer.birthyear,
+                playHand: newPlayer.playHand,
+                playStyle: newPlayer.playStyle,
+                grip: newPlayer.grip,
+                countryId: newPlayer.country.id
             }
         })
     }
@@ -125,11 +144,30 @@ export const createPlayer = async(player: Player) => {
     
 }
 
-export const deletePlayer = async(id: string) => {
-    const prisma = new PrismaClient()
-    return await prisma.player.delete({
-        where: { id: id }
-    })
+export const deletePlayer = async(playerId: string) => {
+    try {
+        const prisma = new PrismaClient()
+        let player = await prisma.player.findFirst({
+            where: { id: playerId }
+        })
+
+        if(player !== null) {
+            return await prisma.player.delete({
+                where: { id: playerId }
+            })
+        }
+
+        //throw 'The given user does not exist'
+    } catch (e) {
+        if (e instanceof Prisma.PrismaClientKnownRequestError) {
+            if (e.code == 'P2025') {
+                console.log("The record to delete does not exist")
+            }
+        }
+
+        throw e
+    }
+    
 }
 
 const playerWithCountry = Prisma.validator<Prisma.PlayerArgs>()({
